+++ pkg/irsa/legacy/legacy.go
@@ -26,7 +26,7 @@ import (
 	"github.com/giantswarm/irsa-operator/pkg/aws/services/iam"
 	"github.com/giantswarm/irsa-operator/pkg/aws/services/route53"
 	"github.com/giantswarm/irsa-operator/pkg/aws/services/s3"
-	"github.com/giantswarm/irsa-operator/pkg/errors"
+	irsaerrors "github.com/giantswarm/irsa-operator/pkg/errors"
 	"github.com/giantswarm/irsa-operator/pkg/key"
 	ctrlmetrics "github.com/giantswarm/irsa-operator/pkg/metrics"
 	"github.com/giantswarm/irsa-operator/pkg/pkcs"
@@ -249,7 +249,7 @@ func (s *Service) Reconcile(ctx context.Context) error {
 		cfConfig := &v1.ConfigMap{}
 		err = s.Client.Get(ctx, types.NamespacedName{Namespace: s.Scope.ClusterNamespace(), Name: s.Scope.ConfigName()}, cfConfig)
 		if apierrors.IsNotFound(err) {
-			if err := errors.IsEmptyCloudfrontDistribution(distribution); err != nil {
+			if err := irsaerrors.IsEmptyCloudfrontDistribution(distribution); err != nil {
 				ctrlmetrics.Errors.WithLabelValues(s.Scope.Installation(), s.Scope.AccountID(), s.Scope.ClusterName(), s.Scope.ClusterNamespace()).Inc()
 				s.Scope.Logger().Error(err, "cloudfront distribution cannot be nil")
 				return err
@@ -336,9 +336,14 @@ func (s *Service) Reconcile(ctx context.Context) error {
 
 	createOIDCProvider := func() error {
 		var identityProviderURLs []string
+		var identityProviderURLsToDelete []string
 		s3Endpoint := fmt.Sprintf("s3.%s.%s", s.Scope.Region(), key.AWSEndpoint(s.Scope.Region()))
 		if (key.IsV18Release(s.Scope.Release()) && !key.IsChina(s.Scope.Region())) || (s.Scope.MigrationNeeded() && !key.IsChina(s.Scope.Region())) {
-			identityProviderURLs = append(identityProviderURLs, util.EnsureHTTPS(cfDomain))
+			if s.Scope.KeepCloudFrontOIDCProvider() {
+				identityProviderURLs = append(identityProviderURLs, util.EnsureHTTPS(cfDomain))
+			} else {
+				identityProviderURLsToDelete = append(identityProviderURLs, util.EnsureHTTPS(cfDomain))
+			}
 		} else {
 			identityProviderURLs = append(identityProviderURLs, util.EnsureHTTPS(fmt.Sprintf("%s/%s", s3Endpoint, s.Scope.BucketName())))
 		}
@@ -347,7 +352,7 @@ func (s *Service) Reconcile(ctx context.Context) error {
 			identityProviderURLs = append(identityProviderURLs, util.EnsureHTTPS(*alias))
 		}
 
-		return s.IAM.EnsureOIDCProviders(identityProviderURLs, key.STSUrl(s.Scope.Region()), customerTags)
+		return s.IAM.EnsureOIDCProviders(identityProviderURLs, identityProviderURLsToDelete, key.STSUrl(s.Scope.Region()), customerTags)
 	}
 	n := func(err error, d time.Duration) {
 		s.Scope.Logger().Info("level", "warning", "message", fmt.Sprintf("retrying backoff in '%s' due to error", d.String()), "stack", fmt.Sprintf("%#v", err))